speech/requirements.txt
fastapi
uvicorn
openai-whisper
TTS==0.22.0
torch==2.1.0
torchaudio==2.1.0
transformers==4.36.2
soundfile==0.12.1
langdetect==1.0.9
python-multipart